Crud Departement

+++ back/src/main/java/com/example/springproject/Repository/DepartementRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DepartementRepository extends CrudRepository<Departement,Long> {
+public interface DepartementRepository extends CrudRepository<Departement,Integer> {
 }
